Add jvm cpp distinction by platforms
