+++ src/jni/wrapper.h
@@ -6,7 +6,11 @@
 #include "exceptions.h"
 #include "init_args.h"
 #include "jvm.h"
+
+#ifndef __ANDROID__
 #include "jvm_loader.h"
+#endif
+
 #include "types.h"
 #include "local_frame.h"
 #include "java_method_signature.h"
